requirements.txt
google-generativeai
PyPDF2
python-pptx
streamlit
fpdf
